+++ memory/src/main/scala/filodb.memory/Block.scala
@@ -5,7 +5,6 @@ import java.util.ConcurrentModificationException
 import java.util.concurrent.atomic.{AtomicBoolean, AtomicReference}
 
 import com.typesafe.scalalogging.StrictLogging
-import org.joda.time.DateTime
 
 /*
 * Useful to establish thread ownership of a buffer.
@@ -203,42 +202,4 @@ class Block(val address: Long, val capacity: Long, val reclaimListener: ReclaimL
   // debug method to set memory to specific value for testing
   private[memory] def set(value: Byte): Unit =
     UnsafeUtils.unsafe.setMemory(address, capacity, value)
-}
-
-/**
- * BlockDetective has utilities for recovering blocks and their reclaim histories, given a corrupt pointer.
- * This helps to track down ownership and causes for corruption issues.
- */
-object BlockDetective {
-  import collection.JavaConverters._
-
-  def containsPtr(ptr: BinaryRegion.NativePointer, blocks: Seq[Block]): Seq[Block] =
-    blocks.filter { blk => ptr >= blk.address && ptr < (blk.address + blk.capacity) }
-
-  def containsPtr(ptr: BinaryRegion.NativePointer, blocks: java.util.List[Block]): Seq[Block] =
-    containsPtr(ptr, blocks.asScala)
-
-  /**
-   * Produces a string report containing reclaim history and ownership changes for
-   * blocks containing a given pointer.
-   * Reclaim history is limited by MaxReclaimLogSize above, thus
-   * the bet is that corruption happens soon after reclaim events.
-   */
-  def stringReport(ptr: BinaryRegion.NativePointer,
-                   manager: PageAlignedBlockManager,
-                   pool: BlockMemFactoryPool): String = {
-    val reclaimEvents = manager.reclaimEventsForPtr(ptr)
-    val timeBucketBlocks = manager.timeBlocksForPtr(ptr)
-    val flushBlocks = pool.blocksContainingPtr(ptr)
-
-    f"=== BlockDetective Report for 0x$ptr%016x ===\nReclaim Events:\n" +
-    reclaimEvents.map { case ReclaimEvent(blk, reclaimTime, oldOwner, remaining) =>
-      f"  Block 0x${blk.address}%016x at ${(new DateTime(reclaimTime)).toString()}%s with $remaining%d bytes left" +
-      oldOwner.map { bmf => s"\tfrom ${bmf.debugString}" }.getOrElse("")
-    }.mkString("\n") +
-    "Time bucketed blocks:\n" +
-    timeBucketBlocks.map(_.debugString).mkString("\n") +
-    "Flush block lists:\n" +
-    flushBlocks.map(_.debugString).mkString("\n")
-  }
 }
