+++ memory/src/main/scala/filodb.memory/MemFactory.scala
@@ -197,7 +197,7 @@ class ArrayBackedMemFactory extends MemFactory {
   *                   block is full.
   * @param bucketTime the timebucket (timestamp) from which to allocate block(s), or None for the general list
   * @param metadataAllocSize the additional size in bytes to ensure is free for writing metadata, per chunk
-  * @param context a set of keys/values for debugging context
+  * @param context a set of keys/values to identify the purpose of this MemFactory for debugging
   * @param markFullBlocksAsReclaimable Immediately mark and fully used block as reclaimable.
   *                                    Typically true during on-demand paging of optimized chunks from persistent store
   */
